+++ test/td_cache/domain_cache_test.exs
@@ -6,7 +6,6 @@ defmodule TdCache.DomainCacheTest do
 
   alias TdCache.DomainCache
   alias TdCache.Redix
-  alias TdCache.Redix.Stream
 
   doctest TdCache.DomainCache
 
@@ -25,33 +24,19 @@ defmodule TdCache.DomainCacheTest do
 
   describe "DomainCache" do
     test "writes a domain entry in redis and reads it back", %{domain: %{id: id} = domain} do
-      {:ok, [3, 1, 1, 1, 1, 0]} = DomainCache.put(domain)
+      {:ok, [0, 4, 1, 1, 1, 1, 0]} = DomainCache.put(domain)
       {:ok, d} = DomainCache.get(id)
-      assert_structs_equal(d, domain, [:id, :name, :external_id])
-    end
-
-    test "publishes an event when a domain is created", %{domain: domain} do
-      {:ok, _} = DomainCache.put(domain)
-      assert {:ok, events} = Stream.read(:redix, ["domain:events"], transform: true)
-      assert [%{event: "domain_created"}] = events
-    end
-
-    test "does not publish an event if publish: false is specified", %{domain: domain} do
-      {:ok, _} = DomainCache.put(domain, publish: false)
-      assert {:ok, []} = Stream.read(:redix, ["domain:events"], transform: true)
+      assert_structs_equal(d, domain, [:id, :name, :external_id, :parent_id])
     end
 
     test "updates a domain entry only if changed", %{domain: domain} do
-      assert {:ok, [3, 1, 1, 1, 1, 0]} = DomainCache.put(domain)
+      assert {:ok, [0, 4, 1, 1, 1, 1, 0]} = DomainCache.put(domain)
       assert {:ok, []} = DomainCache.put(domain)
 
       updated = %{domain | name: "updated name", updated_at: DateTime.utc_now()}
 
-      assert {:ok, [0, 0, 0, 0, 0, 0]} = DomainCache.put(updated)
+      assert {:ok, [1, 4, 0, 0, 0, 0, 0]} = DomainCache.put(updated)
       assert {:ok, %{name: "updated name"}} = DomainCache.get(domain.id)
-
-      assert {:ok, events} = Stream.read(:redix, ["domain:events"], transform: true)
-      assert [%{event: "domain_created"}, %{event: "domain_updated"}] = events
     end
 
     test "deletes an entry in redis", %{domain: %{id: id} = domain} do
@@ -112,7 +97,7 @@ defmodule TdCache.DomainCacheTest do
       children = Enum.flat_map(domains, fn %{id: id} -> build_many(3, parent_id: id) end)
 
       for domain <- children ++ domains ++ parents do
-        DomainCache.put(domain, publish: false)
+        DomainCache.put(domain)
       end
 
       [parents: parents, domains: domains, children: children]